requirements.txt
colorama==0.4.1
requests==2.22.0
beautifulsoup4==4.8.0
python_dateutil==2.8.0
termcolor==1.1.0
PyYAML==5.1.2
